feat: add gate on user delete

+++ backend/app/Http/Controllers/API/v1/Admin/Entity/User/UserController.php
@@ -107,6 +107,10 @@ class UserController extends Controller
      */
     public function destroy(User $user): JsonResponse
     {
+        if(Gate::denies('delete', $user)) {
+            return response()->json(['message' => 'You are not allowed to access this action'], 403);
+        }
+
         return $this->service->delete($user);
     }
 }
